+++ tests/Unit/ProtoTests/CustomLro/out/src/resources/custom_lro_descriptor_config.php
@@ -8,6 +8,7 @@ return [
                     'additionalArgumentMethods' => [
                         'getProject',
                         'getRegion',
+                        'getFoo',
                     ],
                     'getOperationMethod' => 'get',
                     'cancelOperationMethod' => 'cancel',