+++ tests/Unit/ProtoTests/CustomLro/out/src/resources/custom_lro_operations_rest_client_config.php
@@ -17,6 +17,7 @@ return [
                 'queryParams' => [
                     'project',
                     'region',
+                    'foo',
                 ],
             ],
         ],